refactor: split class gamebase
