--- src/application/EndPanel.fxml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.GameResPanel">
-   <center>
-      <VBox alignment="CENTER" prefHeight="200.0" prefWidth="100.0" BorderPane.alignment="CENTER">
-         <children>
-            <Label fx:id="winLabel" contentDisplay="CENTER" text="White Win">
-               <font>
-                  <Font name="Arial Black" size="58.0" />
-               </font>
-            </Label>
-            <Label text="Back To Menu">
-               <font>
-                  <Font name="Arial Black" size="16.0" />
-               </font>
-            </Label>
-            <Button fx:id="BackToMenuButton" mnemonicParsing="false" onAction="#BackToMenuButtonPressed" text="Menu">
-               <VBox.margin>
-                  <Insets top="10.0" />
-               </VBox.margin>
-            </Button>
-         </children>
-      </VBox>
-   </center>
-</BorderPane>
