+++ src/application/Tile.java
@@ -3,6 +3,8 @@ package application;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Rectangle;
 
+import static application.GameBase.TILE_SIZE;
+
 public class Tile extends Rectangle {
 
     private Piece piece;
@@ -20,10 +22,10 @@ public class Tile extends Rectangle {
     }
 
     public Tile(boolean light, int x, int y) {
-        setWidth(CheckersApp.TILE_SIZE);
-        setHeight(CheckersApp.TILE_SIZE);
+        setWidth(TILE_SIZE);
+        setHeight(TILE_SIZE);
 
-        relocate(x * CheckersApp.TILE_SIZE, y * CheckersApp.TILE_SIZE);
+        relocate(x * TILE_SIZE, y * TILE_SIZE);
 
         setFill(light ? Color.valueOf("#feb") : Color.valueOf("#582"));
     }
